Cleanup.

+++ src/common/memory/freelist.h
@@ -33,8 +33,8 @@ struct Freelist
 
 struct Freelist_Alloc_Header
 {
-    size_t size;
-    size_t padding;
+    u64 size;
+    u64 padding;
 };
 
 NAPI void freelist_init(Freelist* freelist, Arena arena);
@@ -45,7 +45,7 @@ NAPI Freelist_Node* freelist_grow(Freelist* freelist, s64 min_increase, Freelist
 NAPI void freelist_insert(Freelist* freelist, Freelist_Node* insert_after, Freelist_Node* node);
 NAPI void freelist_remove(Freelist* freelist, Freelist_Node* prev, Freelist_Node* node);
 
-NAPI Freelist_Node* freelist_find_first(Freelist* freelist, size_t size, size_t align, size_t* padding_out, Freelist_Node** prev_node_out);
+NAPI Freelist_Node* freelist_find_first(Freelist* freelist, s64 size, s64 align, s64* padding_out, Freelist_Node** prev_node_out);
 NAPI void* freelist_allocate(Freelist* freelist, s64 size, s64 align);
 NAPI void freelist_release(Freelist* freelist, void* ptr);
